Moved automation settings to AutoPlayTab and renamed it to AutomationTab (#11765)
* Moved automation settings to AutoPlayTab

* Renamed the AutoPlayTab to AutomationTab

* Added a translation
