+++ android/assets/jsons/translations/template.properties
@@ -890,6 +890,9 @@ Hide =
 HIGHLY EXPERIMENTAL - YOU HAVE BEEN WARNED! = 
 You need to restart the game for this change to take effect. = 
 
+# AutomationTab
+Automation = 
+
 # AutoPlay
 AutoPlay = 
 Show AutoPlay button =
