+++ core/src/com/unciv/ui/popups/options/OptionsPopup.kt
@@ -85,8 +85,8 @@ class OptionsPopup(
             ImageGetter.getImage("OtherIcons/Options"), 24f
         )
         tabs.addPage(
-            "AutoPlay",
-            autoPlayTab(this),
+            "Automation",
+            automationTab(this),
             ImageGetter.getImage("OtherIcons/NationSwap"), 24f
         )
         tabs.addPage(
